+++ src/main/java/chess/controller/ChessController.java
@@ -13,7 +13,6 @@ import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -24,6 +23,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class ChessController {
+
     private final ChessService chessService;
 
     public ChessController(ChessService chessService) {
@@ -47,8 +47,10 @@ public class ChessController {
     }
 
     @PatchMapping(value = "/board/{roomId}")
-    public ResponseEntity<GameStateDto> move(@PathVariable Long roomId, @RequestBody MoveDto moveDto) {
-        return ResponseEntity.ok().body(chessService.move(roomId, moveDto.getSource(), moveDto.getDestination()));
+    public ResponseEntity<GameStateDto> move(@PathVariable Long roomId,
+        @RequestBody MoveDto moveDto) {
+        return ResponseEntity.ok()
+            .body(chessService.move(roomId, moveDto.getSource(), moveDto.getDestination()));
     }
 
     @GetMapping("/rooms/{roomId}/score")
@@ -72,19 +74,8 @@ public class ChessController {
     }
 
     @DeleteMapping("/rooms")
-    public ResponseEntity delete(@RequestBody DeleteRoomDto deleteDto) {
+    public ResponseEntity<String> delete(@RequestBody DeleteRoomDto deleteDto) {
         chessService.deleteBy(deleteDto.getRoomId(), deleteDto.getPassword());
         return ResponseEntity.ok().body("방이 삭제되었습니다.");
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> IllegalArgumentExceptionHandle(RuntimeException exception) {
-        return ResponseEntity.badRequest().body(exception.getMessage());
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> ExceptionHandle(Exception exception) {
-        System.out.println(exception.getMessage());
-        return ResponseEntity.badRequest().body(exception.getMessage());
     }
 }
