+++ src/main/java/chess/controller/ControllerAdvice.java
@@ -1,0 +1,24 @@
+package chess.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+@RestControllerAdvice
+public class ControllerAdvice {
+
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<String> IllegalArgumentExceptionHandle(RuntimeException exception) {
+        return ResponseEntity.badRequest().body(exception.getMessage());
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<String> ExceptionHandle(Exception exception) {
+        logger.error(exception.getMessage());
+        return ResponseEntity.badRequest().body("서버에 문제가 발생했습니다.");
+    }
+}
